Revert "[launcher] Merge upstream/tdx_rtmr (#513)" (#516)
This reverts commit 86a7e85745f6f969bd71270c9734190b8afc2e01.
