+++ client/attest_network_test.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/google/go-tpm-tools/internal"
 	"github.com/google/go-tpm-tools/internal/test"
 	pb "github.com/google/go-tpm-tools/proto/attest"
 	"google.golang.org/protobuf/proto"
@@ -25,7 +24,9 @@ func TestNetworkFetchIssuingCertificate(t *testing.T) {
 		t.Fatalf("Error parsing AK Cert: %v", err)
 	}
 
-	certChain, err := internal.GetCertificateChain(akCert, externalClient)
+	key := &Key{cert: akCert}
+
+	certChain, err := key.getCertificateChain(externalClient)
 	if err != nil {
 		t.Error(err)
 	}
