+++ launcher/image/cloudbuild.yaml
@@ -2,9 +2,9 @@ substitutions:
   '_BASE_IMAGE': ''
   '_OUTPUT_IMAGE_NAME': ''
   '_OUTPUT_IMAGE_FAMILY': ''
-  '_BASE_IMAGE_PROJECT': ''
+  '_BASE_IMAGE_PROJECT': 'cos-cloud'
   '_IMAGE_ENV': ''
-  '_BUCKET_NAME': ''
+  '_BUCKET_NAME': '${PROJECT_ID}_cloudbuild'
   '_CS_LICENSE': ''
   '_SHORT_SHA': ''
