+++ verifier/client.go
@@ -37,24 +37,11 @@ type TokenOptions struct {
 // Challenge from CreateChallenge, optional GcpCredentials linked to the
 // attestation, the Attestation generated from the TPM, and optional container image signatures associated with the workload.
 type VerifyAttestationRequest struct {
-	Challenge      *Challenge
-	GcpCredentials [][]byte
-	// Attestation is for TPM attestation
+	Challenge                *Challenge
+	GcpCredentials           [][]byte
 	Attestation              *attestpb.Attestation
 	ContainerImageSignatures []oci.Signature
 	TokenOptions             TokenOptions
-	// TDCCELAttestation is for TDX CCEL RTMR attestation
-	TDCCELAttestation *TDCCELAttestation
-}
-
-type TDCCELAttestation struct {
-	CcelAcpiTable     []byte
-	CcelData          []byte
-	CanonicalEventLog []byte
-	TdQuote           []byte
-	// still needs following two for GCE info
-	AkCert            []byte
-	IntermediateCerts [][]byte
 }
 
 // VerifyAttestationResponse is the response from a successful
